pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "slovo-dnya"
version = "0.2.0"
description = "Ukrainian word-of-the-day engine powered by music lyrics"
requires-python = ">=3.11"
dependencies = [
    "typer[all]>=0.12",
    "pymongo[srv]>=4.7",
    "python-dotenv>=1.0",
    "stanza>=1.8",
    "deep-translator>=1.11",
    "rich>=13.7",
    "requests>=2.31",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "mongomock>=4.1",
    "ruff>=0.4",
]

[project.scripts]
slovo = "slovo.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["slovo*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = ["E501"]

[tool.coverage.run]
source = ["slovo"]
omit = ["tests/*"]
